create Concerns::TwitterUser::Builder and separate builder methods of TwitterUser
