+++ app/models/twitter_user.rb
@@ -48,8 +48,12 @@ class TwitterUser < ActiveRecord::Base
   include Concerns::TwitterUser::UserInfoAccessor
   include Concerns::TwitterUser::Validation
   include Concerns::TwitterUser::Equalizer
+  include Concerns::TwitterUser::Builder
 
   def __uid_i
+    ActiveSupport::Deprecation.warn(<<-MESSAGE.strip_heredoc)
+      `TwitterUser#__uid_i` is deprecated.
+    MESSAGE
     uid.to_i
   end
 
@@ -83,110 +87,6 @@ class TwitterUser < ActiveRecord::Base
     diffs << "friends(#{self.friend_uids.size},#{tu.friend_uids.size})" if self.friend_uids != tu.friend_uids
     diffs << "followers(#{self.follower_uids.size},#{tu.follower_uids.size})" if self.follower_uids != tu.follower_uids
     diffs
-  end
-
-  def self.build_by_user(user, attrs = {})
-    build_relation = attrs.has_key?(:build_relation) ? attrs.delete(:build_relation) : false
-    tu = new(attrs) do |tu|
-      tu.uid = user.id
-      tu.screen_name = user.screen_name
-      tu.user_info = user.slice(*PROFILE_SAVE_KEYS).to_json # TODO check the type of keys and values
-    end
-    tu.build_relations if build_relation
-    tu
-  end
-
-  def self.build_by_client(client, user, attrs = {})
-    build_by_user(client.user(user), attrs.merge(client: client))
-  end
-
-  def self.build(client, user, option = {})
-    build_by_client(client, user, option)
-  end
-
-  def build_relations
-    uid_i = uid.to_i
-    search_query = "@#{screen_name}"
-
-    if ego_surfing?
-      candidates = [
-        {method: :friends_parallelly, args: [uid_i]},
-        {method: :followers_parallelly, args: [uid_i]},
-        {method: :user_timeline, args: [uid_i]}, # for replying
-        {method: :search, args: [search_query]}, # for replied
-        {method: :home_timeline, args: [uid_i]},
-        {method: :mentions_timeline, args: [uid_i]}, # for replied
-        {method: :favorites, args: [uid_i]} # for favoriting
-      ]
-      if without_friends
-        candidates = candidates.slice(2, candidates.size - 2)
-        _statuses, _search_results, _, _mentions, _favorites = client._fetch_parallelly(candidates)
-        _friends = _followers = []
-      else
-        _friends, _followers, _statuses, _search_results, _, _mentions, _favorites = client._fetch_parallelly(candidates)
-      end
-    else
-      candidates = [
-        {method: :friends_parallelly, args: [uid_i]},
-        {method: :followers_parallelly, args: [uid_i]},
-        {method: :user_timeline, args: [uid_i]},
-        {method: :search, args: [search_query]},
-        {method: :favorites, args: [uid_i]}
-      ]
-      if without_friends
-        candidates = candidates.slice(2, candidates.size - 2)
-        _statuses, _search_results, _favorites = client._fetch_parallelly(candidates)
-        _friends = _followers = []
-      else
-        _friends, _followers, _statuses, _search_results, _favorites = client._fetch_parallelly(candidates)
-      end
-      _mentions = []
-    end
-
-    # Not using uniq for mentions, search_results and favorites intentionally
-
-    client._fetch_parallelly([
-                              {method: :replying, args: [uid_i]}
-                            ])
-
-    _friends.each do |f|
-      friends.build(uid: f.id,
-                       screen_name: f.screen_name,
-                       user_info: f.slice(*PROFILE_SAVE_KEYS).to_json)
-    end
-
-    _followers.each do |f|
-      followers.build(uid: f.id,
-                         screen_name: f.screen_name,
-                         user_info: f.slice(*PROFILE_SAVE_KEYS).to_json)
-    end
-
-    _statuses.each do |s|
-      statuses.build(uid: uid,
-                        screen_name: screen_name,
-                        status_info: s.slice(*Status::STATUS_SAVE_KEYS).to_json)
-    end
-
-    _mentions.each do |m|
-      mentions.build(uid: m.user.id,
-                        screen_name: m.user.screen_name,
-                        status_info: m.slice(*Status::STATUS_SAVE_KEYS).to_json)
-    end
-
-    _search_results.each do |sr|
-      search_results.build(uid: sr.user.id,
-                              screen_name: sr.user.screen_name,
-                              status_info: sr.slice(*Status::STATUS_SAVE_KEYS).to_json,
-                              query: search_query)
-    end
-
-    _favorites.each do |f|
-      favorites.build(uid: f.user.id,
-                         screen_name: f.user.screen_name,
-                         status_info: f.slice(*Status::STATUS_SAVE_KEYS).to_json)
-    end
-
-    true
   end
 
   def save_with_bulk_insert(validate = true)
